@@ -1,70 +1,45 @@
 {
   "name": "create-librarydata",
-  "version": "0.1.0",
-  "private": false,
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "deploy": "gh-pages -d dist",
-    "predeploy": "vue-cli-service build",
-    "start": "vue-cli-service lint --fix && vue-cli-service serve"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "deploy": "gh-pages -d build",
+    "predeploy": "vite build"
   },
   "dependencies": {
     "@turf/bbox": "^6.5.0",
-    "axios": "^1.6.7",
+    "axios": "^1.6.8",
     "colorbrewer": "^1.5.6",
-    "core-js": "^3.36.0",
     "file-saver": "^2.0.5",
-    "jose": "^4.15.4",
-    "maplibre-gl": "^2.4.0",
-    "mitt": "^3.0.1",
+    "jose": "^5.2.4",
+    "maplibre-gl": "^3.6.2",
     "moment": "^2.30.1",
     "papaparse": "^5.4.1",
-    "vue": "^3.2.47",
-    "vue-carbonbadge": "^1.2.1",
-    "vue-maplibre-gl": "^2.1.0",
+    "vue": "^3.4.21",
+    "vue-carbonbadge": "^2.0.0",
+    "vue-maplibre-gl": "^3.1.3",
     "vue-markdown-render": "^2.1.1",
     "vue-router": "^4.3.0",
-    "vuetify": "^3.5.6",
+    "vuetify": "^3.5.15",
     "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.0",
-    "@babel/eslint-parser": "^7.23.10",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-service": "~5.0.8",
-    "@vue/compiler-sfc": "^3.2.47",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.7.1",
-    "gh-pages": "^3.2.3",
-    "raw-loader": "^4.0.2",
-    "sass": "^1.71.1",
-    "sass-loader": "^13.3.3",
-    "webpack-plugin-vuetify": "^2.0.1"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "@babel/eslint-parser"
-    },
-    "rules": {}
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ]
+    "@vitejs/plugin-vue": "^5.0.4",
+    "eslint": "^8.57.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.24.1",
+    "sass": "^1.74.1",
+    "unplugin-fonts": "^1.1.1",
+    "unplugin-vue-components": "^0.26.0",
+    "unplugin-vue-router": "^0.8.5",
+    "vite": "^5.2.8",
+    "vite-plugin-vuetify": "^2.0.3"
+  }
 }
