@@ -1,19 +1,27 @@
-const config = require("../helpers/config.json");
-import axios from "axios";
-import * as csvHelper from "../helpers/csv";
+import axios from 'axios'
+
+import * as csvHelper from '../helpers/csv'
+
+import * as config from '../helpers/config.json'
 
 export const getSchemaData = async (schemaName, localAuthorityCode) => {
-  const url = `${config.schema_url}${schemaName}/${localAuthorityCode}`;
-  const data = await csvHelper.parseUrl(url);
-  return data;
-};
+  const url = `${config.schema_url}${schemaName}/${localAuthorityCode}`
+  const data = await csvHelper.parseUrl(url)
+  return data
+}
 
-export const saveSchemaFile = async (schemaName, localAuthorityCode, content, loginKey) => {
-  const url = `${config.schema_url}${schemaName}/${localAuthorityCode}`;
+export const saveSchemaFile = async (
+  schemaName,
+  localAuthorityCode,
+  content,
+  loginKey
+) => {
+  const url = `${config.schema_url}${schemaName}/${localAuthorityCode}`
   const result = await axios.put(url, content, {
     headers: {
-      'Authorization': `Bearer ${loginKey}`, 'Content-Type': 'text/csv'
+      Authorization: `Bearer ${loginKey}`,
+      'Content-Type': 'text/csv'
     }
-  });
-  return result.status === 200;
-};
+  })
+  return result.status === 200
+}
