@@ -1,6 +1,9 @@
 <template>
   <v-container>
-    <layout-header title="Postcodes to areas" subtitle="Convert postcode locations to statistical areas" />
+    <layout-header
+      title="Postcode to area"
+      subtitle="Convert postcode locations to statistical areas"
+    />
 
     <v-divider inset color="info" class="my-2"></v-divider>
     <markdown-section :markdownText="mdText" />
@@ -10,76 +13,137 @@
 
     <v-sheet color="grey-lighten-5" rounded elevation="0" class="px-5 py-5">
       <v-alert icon="mdi-numeric-1-circle" class="mb-1" title="Load a file">
-        This tool loads CSV files. If your data isn't in CSV format try <strong>Save as</strong> in your software.
-        The first row should be column headings, one column should contain UK postcodes
+        This tool loads CSV files. If your data isn't in CSV format try
+        <strong>Save as</strong> in your software. The first row should be
+        column headings and one column should contain UK postcodes
       </v-alert>
 
-      <file-upload v-bind:files="files" v-on:change-files="confirmFile($event)" />
+      <file-upload
+        v-bind:files="files"
+        v-on:change-files="confirmFile($event)"
+      />
 
-      <v-alert class="mt-8 mb-4" icon="mdi-numeric-2-circle" title="Select file options">
-        The field headings in your data should be displayed below. Choose which one represents the postcodes. If your data
-        includes counts per postcode, also select the field holding the count</v-alert>
+      <v-alert
+        class="mt-8 mb-4"
+        icon="mdi-numeric-2-circle"
+        title="Select file options"
+      >
+        The field headings in your file are displayed below. Choose which column
+        contains the postcodes. If your data includes counts per postcode, also
+        select the field holding the count</v-alert
+      >
 
-      <v-select color="success" v-model="postcodeColumn" :items="columns" label="Select postcode field" variant="outlined"
-        :disabled="columns.length === 0"></v-select>
+      <v-select
+        color="success"
+        v-model="postcodeColumn"
+        :items="columns"
+        label="Select postcode field"
+        variant="outlined"
+        :disabled="columns.length === 0"
+      ></v-select>
 
-      <v-select color="success" :value="countsColumn" :items="columns.filter((c) => c !== postcodeColumn)" :disabled="postcodeColumn === '' ||
-        columns.filter((c) => c !== postcodeColumn).length === 0
-        " label="Count field (optional)" variant="outlined"></v-select>
+      <v-select
+        color="success"
+        :value="countsColumn"
+        :items="columns.filter(c => c !== postcodeColumn)"
+        :disabled="
+          postcodeColumn === '' ||
+          columns.filter(c => c !== postcodeColumn).length === 0
+        "
+        label="Count field (optional)"
+        variant="outlined"
+      ></v-select>
 
-      <v-btn color="info" variant="tonal" size="large" v-on:click="confirmOptions" :disabled="postcodeColumn === ''"
-        append-icon="mdi-chevron-triple-right">
+      <v-btn
+        color="info"
+        variant="tonal"
+        size="large"
+        v-on:click="confirmOptions"
+        :disabled="postcodeColumn === ''"
+        append-icon="mdi-chevron-triple-right"
+      >
         Convert
       </v-btn>
 
       <v-alert class="mt-8 mb-4" icon="mdi-numeric-3-circle" title="Results">
-        Valid postcodes are changed to their statistical area code. Postcodes that are no longer valid are changed to
-        <strong>Terminated</strong>. Postcodes not recognised are changed to <strong>Unknown</strong>.<br />
+        Valid postcodes are changed to their statistical area code. Postcodes
+        that are no longer valid are changed to
+        <strong>Terminated</strong>. Postcodes not recognised are changed to
+        <strong>Unknown</strong>.<br />
       </v-alert>
 
       <span v-if="endTime !== null">
         <p class="mb-2">
-          {{ "Completed in " + getTimeCompleted() + " seconds" }}
+          {{ 'Completed in ' + getTimeCompleted() + ' seconds' }}
         </p>
 
-        <v-card color="grey-lighten-2" variant="outlined" elevation="0" class="mb-2">
-          <v-data-table no-filter :headers="summaryColumns" :items="summaryData" item-key="total">
+        <v-card
+          color="grey-lighten-2"
+          variant="outlined"
+          elevation="0"
+          class="mb-2"
+        >
+          <v-data-table
+            no-filter
+            :headers="summaryColumns"
+            :items="summaryData"
+            item-key="total"
+          >
             <template #bottom></template>
           </v-data-table>
         </v-card>
 
-        <v-btn color="info" variant="tonal" size="large" v-on:click="downloadConvertedFile"
-          append-icon="mdi-content-save">
+        <v-btn
+          color="info"
+          variant="tonal"
+          size="large"
+          v-on:click="downloadConvertedFile"
+          append-icon="mdi-content-save"
+        >
           Save converted file
         </v-btn>
       </span>
 
-      <v-alert class="mt-8 mb-4" icon="mdi-numeric-4-circle" title="Library membership data">
-        Census geographies are required to publish <a href="https://schema.librarydata.uk/membership"
-          target="_blank">library membership data</a>. This tool has calculated a count of members per statistical area.
-        Select the name of your library service and choose a date the postcodes were extracted from your library
-        management system
+      <v-alert
+        class="mt-8 mb-4"
+        icon="mdi-numeric-4-circle"
+        title="Library membership data"
+      >
+        Census geographies are required to publish
+        <a href="https://schema.librarydata.uk/membership" target="_blank"
+          >library membership data</a
+        >. This tool has calculated a count of members per statistical area.
+        Select the name of your library service and choose a date the postcodes
+        were extracted from your library management system
       </v-alert>
 
       <span v-if="endTime !== null">
-
         <service-select v-on:select="authority = $event" />
 
-        <v-text-field class="mt-4" variant="outlined" v-model="extractDate" label="Count date"
-          prepend-inner-icon="mdi-calendar" type="date"></v-text-field>
+        <v-text-field
+          class="mt-4"
+          variant="outlined"
+          v-model="extractDate"
+          label="Count date"
+          prepend-inner-icon="mdi-calendar"
+          type="date"
+        ></v-text-field>
 
-        <v-btn variant="tonal" color="success" v-on:click="downloadSchemaFile" size="large" append-icon="mdi-content-save"
-          :disabled="authority === '' || extractDate === null">Save membership data</v-btn>
-
+        <v-btn
+          variant="tonal"
+          color="success"
+          v-on:click="downloadSchemaFile"
+          size="large"
+          append-icon="mdi-content-save"
+          :disabled="authority === '' || extractDate === null"
+          >Save membership data</v-btn
+        >
       </span>
-
     </v-sheet>
 
     <v-dialog v-model="loading" persistent width="300">
       <v-card>
-        <v-card-title>
-          Please wait
-        </v-card-title>
+        <v-card-title> Please wait </v-card-title>
         <v-card-text>
           <v-progress-linear indeterminate color="info"></v-progress-linear>
         </v-card-text>
@@ -89,26 +153,26 @@
 </template>
 
 <script>
-import FileUpload from "../components/file-upload";
-import MarkDownData from "../markdown/membershippostcodetolsoa.md";
-import ServiceSelect from '../components/service-select'
+import FileUpload from '../components/FileUpload.vue'
+import MarkDownData from '../markdown/membershippostcodetolsoa.md?raw'
+import ServiceSelect from '../components/ServiceSelect.vue'
 
-import moment from "moment";
+import moment from 'moment'
 
-const config = require("../helpers/config.json");
+import { saveAs } from 'file-saver'
 
-import { saveAs } from "file-saver";
+import * as Papa from 'papaparse'
+import * as csvHelper from '../helpers/csv'
+import * as postcodeHelper from '../helpers/postcode'
 
-import * as Papa from "papaparse";
-import * as csvHelper from "../helpers/csv";
-import * as postcodeHelper from "../helpers/postcode";
+import * as config from '../helpers/config.json'
 
 export default {
-  data() {
+  data () {
     return {
       authority: null,
       columns: [],
-      countsColumn: "",
+      countsColumn: '',
       csvData: [],
       endTime: null,
       extractDate: null,
@@ -117,138 +181,143 @@ export default {
       loading: false,
       lsoasCounted: {},
       mdText: MarkDownData,
-      postcodeColumn: "",
+      postcodeColumn: '',
       startTime: null,
       summaryColumns: [
-        { key: "total", title: "Total" },
-        { key: "converted", title: "Successful" },
-        { key: "terminated", title: "Terminated" },
-        { key: "unknown", title: "Unknown" }
+        { key: 'total', title: 'Total' },
+        { key: 'converted', title: 'Successful' },
+        { key: 'terminated', title: 'Terminated' },
+        { key: 'unknown', title: 'Unknown' }
       ],
       summaryData: [{ total: 0, converted: 0, terminated: 0, unknown: 0 }]
-    };
+    }
   },
   methods: {
     confirmFile: async function (files) {
-      this.files = files;
-      let self = this;
-      self.loading = true;
+      this.files = files
+      const self = this
+      self.loading = true
       if (self.files.length > 0) {
-        const data = await csvHelper.parseFile(self.files[0], false);
-        self.columns = data[0];
-        self.csvData = data;
-        self.loading = false;
+        const data = await csvHelper.parseFile(self.files[0], false)
+        self.columns = data[0]
+        self.csvData = data
+        self.loading = false
       }
     },
     confirmOptions: function () {
-      let self = this;
-      self.startTime = moment();
-      self.loading = true;
+      const self = this
+      self.startTime = moment()
+      self.loading = true
       // First get all the postcodes
-      self.summaryData[0].total = 0;
-      self.summaryData[0].converted = 0;
-      self.summaryData[0].unknown = 0;
-      self.summaryData[0].terminated = 0;
-      self.lsoasCounted = { Unknown: 0 };
-      const postcodeColumnIndex = self.columns.indexOf(self.postcodeColumn);
-      const countsColumnIndex = self.columns.indexOf(this.countsColumn);
+      self.summaryData[0].total = 0
+      self.summaryData[0].converted = 0
+      self.summaryData[0].unknown = 0
+      self.summaryData[0].terminated = 0
+      self.lsoasCounted = { Unknown: 0 }
+      const postcodeColumnIndex = self.columns.indexOf(self.postcodeColumn)
+      const countsColumnIndex = self.columns.indexOf(this.countsColumn)
       const postcodes = self.csvData
-        .map((row) => row[postcodeColumnIndex])
-        .filter((p) => p != null && p != self.postcodeColumn);
-      postcodeHelper.extractLsoaLookupFromPostcodes(postcodes, (postcode_lookup) => {
-        self.csvData.forEach((row, idx) => {
-          const postcode = row[postcodeColumnIndex];
-          const stripped = postcode.replace(/\s/g, "");
-          if (idx === 0 && postcode === self.postcodeColumn) {
-            // This is the header row
-            row[postcodeColumnIndex] = "LSOA";
-          } else {
-            // This is a postcode row
-            this.summaryData[0].total++;
-            let lsoa = "";
-            if (postcode && postcode !== "" && postcode_lookup[stripped]) {
-              lsoa = postcode_lookup[stripped];
-              if (lsoa === "Terminated") {
-                row[postcodeColumnIndex] = "Terminated";
-                this.summaryData[0].terminated++;
-              } else if (lsoa === "Unknown") {
-                row[postcodeColumnIndex] = "Unknown";
-                this.summaryData[0].unknown++;
+        .map(row => row[postcodeColumnIndex])
+        .filter(p => p != null && p !== self.postcodeColumn)
+      postcodeHelper.extractLsoaLookupFromPostcodes(
+        postcodes,
+        postcodeLookup => {
+          self.csvData.forEach((row, idx) => {
+            const postcode = row[postcodeColumnIndex]
+            const stripped = postcode.replace(/\s/g, '')
+            if (idx === 0 && postcode === self.postcodeColumn) {
+              // This is the header row
+              row[postcodeColumnIndex] = 'LSOA'
+            } else {
+              // This is a postcode row
+              this.summaryData[0].total++
+              let lsoa = ''
+              if (postcode && postcode !== '' && postcodeLookup[stripped]) {
+                lsoa = postcodeLookup[stripped]
+                if (lsoa === 'Terminated') {
+                  row[postcodeColumnIndex] = 'Terminated'
+                  this.summaryData[0].terminated++
+                } else if (lsoa === 'Unknown') {
+                  row[postcodeColumnIndex] = 'Unknown'
+                  this.summaryData[0].unknown++
+                } else {
+                  // We have a valid lookup
+                  row[postcodeColumnIndex] = lsoa
+                  this.summaryData[0].converted++
+                }
               } else {
-                // We have a valid lookup
-                row[postcodeColumnIndex] = lsoa;
-                this.summaryData[0].converted++;
+                row[postcodeColumnIndex] = 'Unknown'
+                this.summaryData[0].unknown++
+                lsoa = 'Unknown'
               }
-            } else {
-              row[postcodeColumnIndex] = "Unknown";
-              this.summaryData[0].unknown++;
-              lsoa = "Unknown";
+              if (lsoa === 'Terminated') lsoa = 'Unknown'
+              // Ensure our lookup count has the lsoa key value
+              if (!this.lsoasCounted[lsoa]) this.lsoasCounted[lsoa] = 0
+              if (this.countsColumn !== '') {
+                // if count column use that
+                this.lsoasCounted[lsoa] =
+                  this.lsoasCounted[lsoa] + parseInt(row[countsColumnIndex])
+              } else {
+                // else just add one
+                this.lsoasCounted[lsoa]++
+              }
             }
-            if (lsoa === "Terminated") lsoa = "Unknown";
-            // Ensure our lookup count has the lsoa key value
-            if (!this.lsoasCounted[lsoa]) this.lsoasCounted[lsoa] = 0;
-            if (this.countsColumn !== "") {
-              // if count column use that
-              this.lsoasCounted[lsoa] =
-                this.lsoasCounted[lsoa] + parseInt(row[countsColumnIndex]);
-            } else {
-              // else just add one
-              this.lsoasCounted[lsoa]++;
-            }
-          }
-        });
-        self.active_step = 3;
-        this.loading = false;
-        this.endTime = moment();
-      });
+          })
+          self.active_step = 3
+          this.loading = false
+          this.endTime = moment()
+        }
+      )
     },
     downloadConvertedFile: function () {
-      this.downloadFile("converted.csv", this.csvData);
+      this.downloadFile('converted.csv', this.csvData)
     },
     downloadSchemaFile: async function () {
-      let membershipData = [["Local authority", "Count date", "Area code", "Members"]];
-      const date_string = moment(this.extractDate).format("YYYY-MM-DD");
+      const membershipData = [
+        ['Local authority', 'Count date', 'Area code', 'Members']
+      ]
+      const dateString = moment(this.extractDate).format('YYYY-MM-DD')
       Object.keys(this.lsoasCounted)
-        .filter((k) => k !== "Unknown" && k !== "Terminated")
-        .forEach((lsoa) => {
-          let count = this.lsoasCounted[lsoa];
+        .filter(k => k !== 'Unknown' && k !== 'Terminated')
+        .forEach(lsoa => {
+          const count = this.lsoasCounted[lsoa]
           membershipData.push([
             this.authority.name,
-            date_string,
+            dateString,
             lsoa,
-            count > 4 ? count.toString() : "x"
-          ]);
-        });
-      let count = this.lsoasCounted["Unknown"];
+            count > 4 ? count.toString() : 'x'
+          ])
+        })
+      const count = this.lsoasCounted.Unknown
       // Push unknown and terminated (all as unknown)
-      if (this.lsoasCounted["Unknown"] > 0) {
+      if (this.lsoasCounted.Unknown > 0) {
         membershipData.push([
           this.authority.name,
-          date_string,
-          "Unknown",
-          count > 4 ? count.toString() : "x"
-        ]);
+          dateString,
+          'Unknown',
+          count > 4 ? count.toString() : 'x'
+        ])
       }
-      this.downloadFile("membership.csv", membershipData);
+      this.downloadFile('membership.csv', membershipData)
     },
     downloadFile: function (filename, data) {
-      var blob = new Blob([Papa.unparse(data)], { type: "text/plain;charset=utf-8" });
-      saveAs(blob, filename);
+      const blob = new Blob([Papa.unparse(data)], {
+        type: 'text/plain;charset=utf-8'
+      })
+      saveAs(blob, filename)
     },
     getTimeCompleted: function () {
       if (this.startTime && this.endTime) {
-        return Math.round(
-          moment.duration(this.endTime.diff(this.startTime))
-        );
+        return Math.round(moment.duration(this.endTime.diff(this.startTime)))
       }
-      return "";
+      return ''
     }
   },
   components: {
-    "file-upload": FileUpload,
-    "service-select": ServiceSelect
+    'file-upload': FileUpload,
+    'service-select': ServiceSelect
   }
-};
+}
 </script>
 
 <style scoped></style>
