@@ -3,18 +3,28 @@
     <v-app-bar elevation="0" flat color="grey-lighten-5">
       <v-app-bar-nav-icon class="hidden-lg-and-up" @click="drawer = !drawer" />
       <v-container>
-        <img class="mr-3" :src="require('./assets/logo.png')" height="40" />
+        <img class="mr-3" src="/assets/logo.png" height="40" />
         <v-spacer></v-spacer>
       </v-container>
     </v-app-bar>
     <v-navigation-drawer border v-model="drawer" elevation="0">
       <v-list nav>
-        <v-list-item variant="tonal" color="info" :to="'/'" active-class="highlighted"
-          :class="'/' === $route.path ? 'highlighted' : ''" prepend-icon="mdi-home-circle">
+        <v-list-item
+          color="info"
+          :to="'/'"
+          active-class="highlighted"
+          :class="'/' === $route.path ? 'highlighted' : ''"
+          prepend-icon="mdi-home-circle"
+        >
           <v-list-item-title>Home</v-list-item-title>
         </v-list-item>
-        <v-list-item variant="tonal" color="info" :to="'/about'" active-class="highlighted"
-          :class="'/about' === $route.path ? 'highlighted' : ''" prepend-icon="mdi-information">
+        <v-list-item
+          color="info"
+          :to="'/about'"
+          active-class="highlighted"
+          :class="'/about' === $route.path ? 'highlighted' : ''"
+          prepend-icon="mdi-information"
+        >
           <v-list-item-title>About</v-list-item-title>
         </v-list-item>
         <v-divider />
@@ -22,8 +32,12 @@
           <template v-slot:activator="{ props }">
             <v-list-item v-bind="props" title="Events"></v-list-item>
           </template>
-          <v-list-item variant="tonal" color="success" :to="'/events'" active-class="highlighted"
-            :class="'/events' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/events'"
+            active-class="highlighted"
+            :class="'/events' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Events dataset</v-list-item-title>
           </v-list-item>
         </v-list-group>
@@ -31,12 +45,20 @@
           <template v-slot:activator="{ props }">
             <v-list-item v-bind="props" title="Libraries"></v-list-item>
           </template>
-          <v-list-item variant="tonal" color="success" :to="'/libraries'" active-class="highlighted"
-            :class="'/libraries' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/libraries'"
+            active-class="highlighted"
+            :class="'/libraries' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Libraries dataset</v-list-item-title>
           </v-list-item>
-          <v-list-item variant="tonal" color="success" :to="'/libraries-edit'" active-class="highlighted"
-            :class="'/libraries' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/libraries-edit'"
+            active-class="highlighted"
+            :class="'/libraries' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Edit libraries</v-list-item-title>
           </v-list-item>
         </v-list-group>
@@ -44,8 +66,12 @@
           <template v-slot:activator="{ props }">
             <v-list-item v-bind="props" title="Loans"></v-list-item>
           </template>
-          <v-list-item variant="tonal" color="success" :to="'/loans'" active-class="highlighted"
-            :class="'/loans' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/loans'"
+            active-class="highlighted"
+            :class="'/loans' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Loans dataset</v-list-item-title>
           </v-list-item>
         </v-list-group>
@@ -53,16 +79,28 @@
           <template v-slot:activator="{ props }">
             <v-list-item v-bind="props" title="Membership"></v-list-item>
           </template>
-          <v-list-item variant="tonal" color="success" :to="'/membership'" active-class="highlighted"
-            :class="'/membership' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/membership'"
+            active-class="highlighted"
+            :class="'/membership' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Membership dataset</v-list-item-title>
           </v-list-item>
-          <v-list-item variant="tonal" color="success" :to="'/postcode-to-lsoa'" active-class="highlighted"
-            :class="'/postcode-to-lsoa' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/postcode-to-lsoa'"
+            active-class="highlighted"
+            :class="'/postcode-to-lsoa' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Postcode converter</v-list-item-title>
           </v-list-item>
-          <v-list-item variant="tonal" color="success" :to="'/membership-map'" active-class="highlighted"
-            :class="'/membership-map' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/membership-map'"
+            active-class="highlighted"
+            :class="'/membership-map' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Membership map</v-list-item-title>
           </v-list-item>
         </v-list-group>
@@ -70,8 +108,14 @@
           <template v-slot:activator="{ props }">
             <v-list-item v-bind="props" title="Mobile libraries"></v-list-item>
           </template>
-          <v-list-item variant="tonal" color="success" :to="'/mobile-library-stops'" active-class="highlighted"
-            :class="'/mobile-library-stops' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/mobile-library-stops'"
+            active-class="highlighted"
+            :class="
+              '/mobile-library-stops' === $route.path ? 'highlighted' : ''
+            "
+          >
             <v-list-item-title>Mobile library stops dataset</v-list-item-title>
           </v-list-item>
         </v-list-group>
@@ -79,8 +123,12 @@
           <template v-slot:activator="{ props }">
             <v-list-item v-bind="props" title="Visits"></v-list-item>
           </template>
-          <v-list-item variant="tonal" color="success" :to="'/physical-visits'" active-class="highlighted"
-            :class="'/physical-visits' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/physical-visits'"
+            active-class="highlighted"
+            :class="'/physical-visits' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Physical visits dataset</v-list-item-title>
           </v-list-item>
         </v-list-group>
@@ -88,8 +136,12 @@
           <template v-slot:activator="{ props }">
             <v-list-item v-bind="props" title="Catalogue"></v-list-item>
           </template>
-          <v-list-item variant="tonal" color="success" :to="'/stock-summary'" active-class="highlighted"
-            :class="'/stock-summary' === $route.path ? 'highlighted' : ''">
+          <v-list-item
+            color="success"
+            :to="'/stock-summary'"
+            active-class="highlighted"
+            :class="'/stock-summary' === $route.path ? 'highlighted' : ''"
+          >
             <v-list-item-title>Stock summary dataset</v-list-item-title>
           </v-list-item>
         </v-list-group>
@@ -118,21 +170,21 @@
 </style>
 
 <script>
-import Footer from "./components/layout-footer";
+import Footer from './components/Footer.vue'
 
 export default {
-  name: "App",
+  name: 'App',
   components: {
-    "layout-footer": Footer
+    'layout-footer': Footer
   },
   methods: {
-    async logout() {
-      await this.$store.dispatch("logout");
-      this.$router.push("/");
+    async logout () {
+      await this.$store.dispatch('logout')
+      this.$router.push('/')
     }
   },
   data: () => ({
     drawer: null
   })
-};
+}
 </script>
