@@ -1,18 +1,18 @@
 <template>
-  <v-card elevation="0" color="grey-lighten-5">
+  <v-card elevation="0" color="grey-lighten-5" class="pt-4">
     <v-file-input
       chips
+      show-size
       accept="text/csv"
       counter
       variant="outlined"
       bg-color="white"
       multiple
-      label="Select a file"
+      label="Select a file on your computer"
       :model-value="files"
       v-on:update:model-value="updateSelection($event)"
-      prepend-inner-icon="mdi-file-document"
-      prepend-icon=""
-      hint="This file loader only accepts CSV files"
+      prepend-icon="mdi-file-document"
+      hint="Only files with an extension of .csv will be displayed"
       persistent-hint
     ></v-file-input>
   </v-card>
@@ -20,11 +20,11 @@
 
 <script>
 export default {
-  props: ["files"],
+  props: ['files'],
   methods: {
-    updateSelection(files) {
-      this.$emit("change-files", files);
+    updateSelection (files) {
+      this.$emit('change-files', files)
     }
   }
-};
+}
 </script>
