@@ -4,8 +4,9 @@
     <p class="text-subtitle-1 text-center">{{ subtitle }}</p>
   </v-container>
 </template>
+
 <script>
 export default {
-  props: ["title", "subtitle"]
-};
+  props: ['title', 'subtitle']
+}
 </script>
