@@ -1,13 +1,15 @@
 import { createApp } from 'vue'
 import { createVuetify } from 'vuetify'
+
 import VueMaplibreGl from 'vue-maplibre-gl'
+
 import './styles/main.scss'
 
 import * as labs from 'vuetify/labs/components'
 
-import Header from './components/layout-header'
-import Markdown from './components/markdown-section'
-import UserProfile from './components/user-profile'
+import Header from './components/Header.vue'
+import Markdown from './components/Markdown.vue'
+import UserProfile from './components/UserProfile.vue'
 
 import router from './router'
 
