@@ -2,8 +2,8 @@
 
 import { register } from 'register-service-worker'
 
-if (process.env.NODE_ENV === 'production') {
-  register(`${process.env.BASE_URL}service-worker.js`, {
+if (import.meta.env.NODE_ENV === 'production') {
+  register(`${import.meta.env.BASE_URL}service-worker.js`, {
     ready () {
       console.log(
         'App is being served from cache by a service worker.\n' +
@@ -23,7 +23,9 @@ if (process.env.NODE_ENV === 'production') {
       console.log('New content is available; please refresh.')
     },
     offline () {
-      console.log('No internet connection found. App is running in offline mode.')
+      console.log(
+        'No internet connection found. App is running in offline mode.'
+      )
     },
     error (error) {
       console.error('Error during service worker registration:', error)
