@@ -1,19 +1,23 @@
 <template>
   <v-container>
-    <layout-header title="Accessibility" subtitle="How this site meets accessibility requirements" />
+    <layout-header
+      title="Accessibility"
+      subtitle="How this site meets accessibility requirements"
+    />
     <markdown-section :markdownText="mdText" />
   </v-container>
 </template>
+
 <script>
-import MarkDownData from "../markdown/accessibility.md";
+import MarkDownData from '../markdown/accessibility.md?raw'
 
 export default {
-  data() {
+  data () {
     return {
       mdText: MarkDownData
-    };
+    }
   },
   methods: {},
   components: {}
-};
+}
 </script>
