@@ -1,10 +1,19 @@
 <template>
   <v-container>
-    <layout-header title="Create" subtitle="Helping libraries publish and use library open data" />
+    <layout-header
+      title="Create"
+      subtitle="Helping libraries publish and use library open data"
+    />
     <markdown-section :markdownText="mdText" />
     <v-container class="text-center" v-if="!this.$store.state.loginKey">
-      <v-btn prepend-icon="mdi-login-variant" variant="tonal" size="x-large" color="success" class="mx-1 my-1"
-        :to="'/login'">
+      <v-btn
+        prepend-icon="mdi-login-variant"
+        variant="tonal"
+        size="x-large"
+        color="success"
+        class="mx-1 my-1"
+        :to="'/login'"
+      >
         Log in
       </v-btn>
     </v-container>
@@ -13,16 +22,17 @@
     </div>
   </v-container>
 </template>
+
 <script>
-import MarkDownData from "../markdown/home.md";
+import MarkDownData from '../markdown/home.md?raw'
 
 export default {
-  data() {
+  data () {
     return {
       mdText: MarkDownData
-    };
+    }
   },
   methods: {},
   components: {}
-};
+}
 </script>
