@@ -1,9 +1,18 @@
 <template>
   <v-container>
-    <layout-header title="Libraries" subtitle="Public library locations, hours, and contact details" />
+    <layout-header
+      title="Libraries"
+      subtitle="Public library locations, hours, and contact details"
+    />
     <p class="text-center">
-      <v-btn size="large" prepend-icon="mdi-information-outline" variant="tonal" color="info"
-        href="https://schema.librarydata.uk/libraries" target="_blank">
+      <v-btn
+        size="large"
+        prepend-icon="mdi-information-outline"
+        variant="tonal"
+        color="info"
+        href="https://schema.librarydata.uk/libraries"
+        target="_blank"
+      >
         Explore the locations data schema
       </v-btn>
     </p>
@@ -11,21 +20,19 @@
     <v-divider inset color="info" class="my-2"></v-divider>
     <markdown-section :markdownText="mdText" />
     <v-divider inset color="info" class="my-2"></v-divider>
-
   </v-container>
 </template>
 
 <script>
-import MarkDownData from '../markdown/librariesdata.md'
+import MarkDownData from '../markdown/librariesdata.md?raw'
 
 export default {
-  data() {
+  data () {
     return {
       mdText: MarkDownData
     }
   },
-  methods: {
-  },
+  methods: {},
   components: {}
 }
 </script>
