@@ -2,7 +2,13 @@
   <v-container>
     <layout-header title="Loans" subtitle="Lending of library items" />
     <p class="text-center">
-      <v-btn size="large" variant="tonal" color="info" href="https://schema.librarydata.uk/loans" target="_blank">
+      <v-btn
+        size="large"
+        variant="tonal"
+        color="info"
+        href="https://schema.librarydata.uk/loans"
+        target="_blank"
+      >
         Explore the loans data schema
       </v-btn>
     </p>
@@ -10,20 +16,19 @@
     <v-divider inset color="info" class="my-2"></v-divider>
     <markdown-section :markdownText="mdText" />
     <v-divider inset color="info" class="my-2"></v-divider>
-
   </v-container>
 </template>
 
 <script>
-import MarkDownData from "../markdown/loansdata.md";
+import MarkDownData from '../markdown/loansdata.md?raw'
 
 export default {
-  data() {
+  data () {
     return {
       mdText: MarkDownData
-    };
+    }
   },
   methods: {},
   components: {}
-};
+}
 </script>
