@@ -1,6 +1,9 @@
 <template>
   <v-container>
-    <layout-header title="Profile" subtitle="See access and profile information" />
+    <layout-header
+      title="Profile"
+      subtitle="See access and profile information"
+    />
     <markdown-section :markdownText="mdText" />
     <section>
       <v-container>
@@ -9,20 +12,21 @@
     </section>
   </v-container>
 </template>
+
 <script>
-import MarkDownData from "../markdown/profile.md";
+import MarkDownData from '../markdown/profile.md?raw'
 
 export default {
-  data() {
+  data () {
     return {
       mdText: MarkDownData
-    };
+    }
   },
   methods: {
-    logout() {
-      this.$store.dispatch("logout");
+    logout () {
+      this.$store.dispatch('logout')
     }
   },
   components: {}
-};
+}
 </script>
